# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "tech_check http api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/v1/questions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "questions list",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page",
                        "name": "pagination[page]",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "count",
                        "name": "pagination[count]",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "created_at",
                        "name": "sorts[created_at]",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "description": "updated_at",
                        "name": "sorts[updated_at]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "text",
                        "name": "filters[text]",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "junior",
                            "middle",
                            "senior"
                        ],
                        "type": "string",
                        "description": "grade",
                        "name": "filters[grade]",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.list"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Question"
                                            }
                                        },
                                        "pagination": {
                                            "$ref": "#/definitions/dto.Pagination"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "create question",
                "parameters": [
                    {
                        "description": "question create request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.QuestionCreate"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Question"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/v1/questions/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "get question by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Question"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "questions"
                ],
                "summary": "delete question by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "questions"
                ],
                "summary": "update profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "question id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "question update request",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.QuestionUpdate"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/responses.success"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Question"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "constants.GradeName": {
            "type": "string",
            "enum": [
                "junior",
                "middle",
                "senior"
            ],
            "x-enum-varnames": [
                "GradeJunior",
                "GradeMiddle",
                "GradeSenior"
            ]
        },
        "dto.Pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "page_count": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "models.Question": {
            "type": "object",
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "grade": {
                    "$ref": "#/definitions/constants.GradeName"
                },
                "id": {
                    "type": "string"
                },
                "text": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "requests.QuestionCreate": {
            "type": "object",
            "required": [
                "category_id",
                "grade",
                "text"
            ],
            "properties": {
                "category_id": {
                    "type": "string"
                },
                "grade": {
                    "type": "string",
                    "enum": [
                        "junior",
                        "middle",
                        "senior"
                    ]
                },
                "text": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 3
                }
            }
        },
        "requests.QuestionUpdate": {
            "type": "object",
            "required": [
                "grade",
                "text"
            ],
            "properties": {
                "grade": {
                    "type": "string",
                    "enum": [
                        "junior",
                        "middle",
                        "senior"
                    ]
                },
                "text": {
                    "type": "string",
                    "maxLength": 200,
                    "minLength": 3
                }
            }
        },
...
        "responses.list": {
            "type": "object",
            "properties": {
                "data": {},
                "pagination": {}
            }
        },
        "responses.success": {
            "type": "object",
            "properties": {
                "data": {}
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}